HUD system implemented and Issues solved
Now the HUD is divided in diferent "HUD elements" that corresponds to
the main function of each of the windows of the gui. This way, we can
divide the elements and functionality of each of this parts eithout
having all the gui elements in one unic class, like HUD before. For
example:

- Now we have 2 hudElements:

* hudBelt: in charge of the main hud belt, that has the stamina, life,
mana,  quick skills, quick, inventory, and minipanel of functions.

* hudPause: it's in charge of the pause menu and it's options.

* Other hud elements would be: hudEquipement, hudSkills, hudStats,
hudShop,...

* Now, the j1HUD works like j1Game or j1App, it just iterates the
hud_elements that are stored in a array, calling all their iteration
functions.

- Also, the issue about the player moving when clicking the gui is
solved