All GuiSkills HUD logic implemented
When unlocking an skill at the skill tree a skill will appear at the
skill display at the belt.
Also, you can change the skil being used by clicking on a skill
displayed at the belt. Yay
The only thing is that it still doesn't need blood charges to buy a
skill